Split Behavior group action summary  - Will be required for the behavior group wizard
